requirements.txt
# Core dependencies
langchain==0.1.16
langchain-community==0.0.34
langchain-google-genai==1.0.1
chromadb==0.4.24

# HTML processing
beautifulsoup4==4.12.3
lxml==5.1.0                    # Streaming HTML parsing
trafilatura==1.8.0             # Content extraction
selectolax==0.3.17             # High-speed parsing backup

# Web framework cho API
fastapi==0.110.0
uvicorn==0.29.0

# Data processing
pandas==2.2.1
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# Logging và monitoring
loguru==0.7.2
rich==13.7.1

# Testing (optional nhưng recommended)
pytest==8.1.1
pytest-asyncio==0.23.6

# Development tools
ipython==8.22.2
jupyter==1.0.0  # Để test từng phần code

# Additional for better development experience
tqdm==4.66.2  # Progress bars
httpx==0.27.0  # HTTP client cho API calls
nltk==3.8.1
PyYAML~=6.0.2
requests~=2.32.3